feat : 꿀팁 게시글 상세조회(댓글 조회 미구현) #33

+++ backend/src/test/java/com/gwangjubob/livealone/backend/service/TipServiceTest.java
@@ -7,12 +7,14 @@ import com.gwangjubob.livealone.backend.domain.repository.TipCommentRepository;
 import com.gwangjubob.livealone.backend.domain.repository.TipRepository;
 import com.gwangjubob.livealone.backend.domain.repository.UserRepository;
 import com.gwangjubob.livealone.backend.dto.tip.TipCreateDto;
+import com.gwangjubob.livealone.backend.dto.tip.TipDetailViewDto;
 import com.gwangjubob.livealone.backend.dto.tip.TipUpdateDto;
 import com.gwangjubob.livealone.backend.dto.tip.TipViewDto;
 import com.gwangjubob.livealone.backend.dto.tipcomment.TipCommentCreateDto;
 import com.gwangjubob.livealone.backend.dto.tipcomment.TipCommentUpdateDto;
 import com.gwangjubob.livealone.backend.dto.user.UserInfoDto;
 import com.gwangjubob.livealone.backend.mapper.TipCreateMapper;
+import com.gwangjubob.livealone.backend.mapper.TipDetailViewMapper;
 import com.gwangjubob.livealone.backend.mapper.TipUpdateMapper;
 import com.gwangjubob.livealone.backend.mapper.UserInfoMapper;
 import org.apache.catalina.User;
@@ -33,7 +35,6 @@ import java.util.Optional;
 @ExtendWith(SpringExtension.class)
 @SpringBootTest
 //@Transactional
-@Disabled
 public class TipServiceTest {
     private TipRepository tipRepository;
     private TipService tipService;
@@ -42,11 +43,12 @@ public class TipServiceTest {
     private TipCommentRepository tipCommentRepository;
     private TipCreateMapper tipCreateMapper;
     private TipUpdateMapper tipUpdateMapper;
+    private TipDetailViewMapper tipDetailViewMapper;
 
     @Autowired
     public TipServiceTest(TipCommentService tipCommentService, TipRepository tipRepository, TipService tipService,
                             UserRepository userRepository, TipCommentRepository tipCommentRepository, TipCreateMapper tipCreateMapper,
-                          TipUpdateMapper tipUpdateMapper){
+                          TipUpdateMapper tipUpdateMapper, TipDetailViewMapper tipDetailViewMapper){
         this.tipRepository = tipRepository;
         this.tipService = tipService;
         this.tipCommentService = tipCommentService;
@@ -54,6 +56,7 @@ public class TipServiceTest {
         this.tipCommentRepository = tipCommentRepository;
         this.tipCreateMapper = tipCreateMapper;
         this.tipUpdateMapper = tipUpdateMapper;
+        this.tipDetailViewMapper = tipDetailViewMapper;
     }
 
     @Test
@@ -63,9 +66,9 @@ public class TipServiceTest {
 
         Optional<UserEntity> optionalUser = userRepository.findByNickname(testNickname); // 사용자 정보
 
-        String category = "item";
-        String title = "꿀템테스트333";
-        String content = "게시글테스트1 텍스트";
+        String category = "tip";
+        String title = "꿀팁 제목테스트";
+        String content = "꿀팁 내용 테스트";
         byte[] bannerImg = null;
 
         // when
@@ -103,7 +106,7 @@ public class TipServiceTest {
     @Test
     public void 카테고리별_리스트_조회_테스트(){
         // given
-        String category = "tip";
+        String category = "recipe";
 
         // when
         List<TipViewDto> result = tipService.viewTip(category);
@@ -117,11 +120,19 @@ public class TipServiceTest {
     @Test
     public void 게시글_상세_조회_테스트() {
         // given
-        Integer idx = 22;
+        Integer idx = 36;
 
-        // when
+        Optional<TipEntity> testTip = tipRepository.findByIdx(idx);
 
-        // then
+        if(testTip.isPresent()){
+            // when
+            TipEntity tipEntity = testTip.get();
+            TipDetailViewDto tipDto = tipDetailViewMapper.toDto(tipEntity);
+            tipDto.setUserNickname(tipEntity.getUser().getNickname());
+
+            // then
+            System.out.println(tipDto.toString());
+        }
 
     }
 
@@ -137,46 +148,55 @@ public class TipServiceTest {
         String content = null;
         byte[] bannerImg = null;
 
-        TipEntity tip = tipRepository.findByIdx(testIdx).get(); // idx에 해당하는 게시글 가져오기
-        UserEntity user = userRepository.findByNickname(testNickname).get();
-        // when
-        if(testNickname.equals(tip.getUser().getNickname())){
-            // 로그인 한 닉네임과 글 작성자가 같으면 수정 가능
-            TipUpdateDto updateDto = TipUpdateDto.builder()
-                    .idx(testIdx)
-                    .category(category)
-                    .title(title)
-                    .content(content)
-                    .bannerImg(bannerImg)
-                    .build();
+        Optional<TipEntity> testTip = tipRepository.findByIdx(testIdx);
 
-            TipEntity updateEntity = tipUpdateMapper.toEntity(updateDto);
-            updateEntity.setUser(user);
-            updateEntity.setTime(LocalDateTime.now());
+        if(testTip.isPresent()){
+            TipEntity tip = testTip.get(); // idx에 해당하는 게시글 가져오기
+            UserEntity user = userRepository.findByNickname(testNickname).get();
+            // when
+            if(testNickname.equals(tip.getUser().getNickname())){
+                // 로그인 한 닉네임과 글 작성자가 같으면 수정 가능
+                TipUpdateDto updateDto = TipUpdateDto.builder()
+                        .idx(testIdx)
+                        .category(category)
+                        .title(title)
+                        .content(content)
+                        .bannerImg(bannerImg)
+                        .build();
 
-            tipRepository.save(updateEntity);
+                TipEntity updateEntity = tipUpdateMapper.toEntity(updateDto);
+                updateEntity.setUser(user);
+                updateEntity.setUpdateTime(LocalDateTime.now());
+
+                tipRepository.save(updateEntity);
+            }
         }
 
         // then
-//        Optional<TipEntity> result = tipRepository.findByTitle(title);
-//        if(result.isPresent()) {
-//            TipEntity entity = result.get();
-//            System.out.println(entity.toString());
-//        }
+        Optional<TipEntity> result = tipRepository.findByTitle(title);
+        if(result.isPresent()) {
+            TipEntity entity = result.get();
+            System.out.println(entity.toString());
+        }
     }
 
     @Test
     public void 게시글_삭제_테스트(){
-        String testId = "ssafy";
-        Integer idx = 22;
+        // given
+        String testId = "test";
+        Integer idx = 37;
 
-        TipEntity tip = tipRepository.findByIdx(idx).get();
-        // when
-        if(testId.equals(tip.getUser().getId())){
-            // 작성자와 로그인한 사용자의 아이디가 같다면 삭제
-            // 삭제 시 댓/대댓글도 모두 삭제
-            // cascade 걸려있으니 그냥 삭제?
-            tipRepository.delete(tip);
+        Optional<TipEntity> testTip = tipRepository.findByIdx(idx);
+
+        if(testTip.isPresent()){
+            TipEntity tip = testTip.get();
+            // when
+            if(testId.equals(tip.getUser().getId())){
+                // 작성자와 로그인한 사용자의 아이디가 같다면 삭제
+                // 삭제 시 댓/대댓글도 모두 삭제
+                // cascade 걸려있으니 그냥 삭제?
+                tipRepository.delete(tip);
+            }
         }
     }
 
@@ -210,7 +230,6 @@ public class TipServiceTest {
         String nickname = "비밀번호는 ssafy 입니다.";
         String content = "대댓글 테스트";
         byte[] bannerImg = null;
-//        String bannerImg = null;
         Integer idx = 19;
         Integer postIdx = 22;
         Integer upIdx = 19;
